results/permissions.py
from rest_framework import permissions

class IsAdminOrSuperUser(permissions.BasePermission):
    def has_permission(self, request, view):
        return request.user.role_id == 1 or request.user.role_id == 2