requirements.txt
absl-py==0.9.0
astor==0.8.1
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
Django==2.2.6
django-ajax-selects==1.7.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
gast==0.2.2
google==2.0.3
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
httplib2==0.15.0
idna==2.8
imageio==2.6.1
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
networkx==2.4
numpy==1.17.3
oauthlib==3.1.0
opencv-python==4.1.1.26
opt-einsum==3.1.0
Pillow==6.2.1
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.4.1
six==1.13.0
sklearn==0.0
soupsieve==1.9.5
sqlparse==0.3.0
tb-nightly==1.14.0a20190603
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
uritemplate==3.0.1
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
gunicorn==20.0.4
django-cors-headers==3.2.0
